feat(be): enable support role to delete attachments in communications module

See merge request camac-ng/camac-ng!10742

+++ django/camac/communications/views.py
@@ -1,6 +1,6 @@
 from django.conf import settings
 from django.db import transaction
-from django.db.models import Exists, OuterRef
+from django.db.models import Exists, OuterRef, Value
 from django.http import FileResponse
 from django.urls import reverse
 from django.utils.translation import gettext
@@ -8,7 +8,12 @@ from django_presigned_url.presign_urls import verify_presigned_request
 from rest_framework import response
 from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound, PermissionDenied, ValidationError
-from rest_framework.mixins import CreateModelMixin, ListModelMixin, RetrieveModelMixin
+from rest_framework.mixins import (
+    CreateModelMixin,
+    DestroyModelMixin,
+    ListModelMixin,
+    RetrieveModelMixin,
+)
 from rest_framework.viewsets import GenericViewSet
 from rest_framework_json_api.views import (
     AutoPrefetchMixin,
@@ -18,6 +23,8 @@ from rest_framework_json_api.views import (
 )
 
 from camac.instance.mixins import InstanceQuerysetMixin
+from camac.user.permissions import permission_aware
+from camac.utils import is_support
 
 from . import filters, models, serializers
 
@@ -25,12 +32,19 @@ from . import filters, models, serializers
 class InvolvedInTopicQuerysetMixin:
     def get_queryset(self, *args, **kwargs):
         """Limit queryset to only contain entities where user is involved in topic."""
-        my_entity = models.entity_for_current_user(self.request)
         qs = super().get_queryset(*args, **kwargs)
+
+        if is_support(self.request):
+            return qs
 
         # Now we additionally limit entities to only list things that
         # belong to topics that I'm actually involved in
         entity_field = self.instance_field.replace("instance", "involved_entities")
+        my_entity = models.entity_for_current_user(self.request)
+
+        if not my_entity:  # pragma: no cover
+            return qs.none()
+
         qs = qs.filter(**{f"{entity_field}__contains": [my_entity]})
 
         return qs
@@ -45,8 +59,19 @@ class TopicView(InvolvedInTopicQuerysetMixin, InstanceQuerysetMixin, ModelViewSe
     queryset = models.CommunicationsTopic.objects
 
     def _annotate_has_unread(self, qs):
+        """Annotate the given queryset, so it has a `has_unread` boolean flag.
+
+        For support roles, or non-communication-module entity users, this will
+        always be `True`.
+        """
+
+        my_entity = models.entity_for_current_user(self.request)
+
+        if is_support(self.request) or not my_entity:
+            return qs.annotate(has_unread=Value(True))
+
         unread_messages = models.CommunicationsMessage.objects.all().exclude(
-            read_by__entity=models.entity_for_current_user(self.request)
+            read_by__entity=my_entity
         )
         qs_out = qs.annotate(
             has_unread=Exists(unread_messages.filter(topic=OuterRef("pk")))
@@ -63,6 +88,13 @@ class TopicView(InvolvedInTopicQuerysetMixin, InstanceQuerysetMixin, ModelViewSe
         qs = self._annotate_has_unread(qs)
         qs = self._annotate_dossier_number(qs)
         return qs
+
+    @permission_aware
+    def has_create_permission(self, *args, **kwargs):
+        return True
+
+    def has_create_permission_for_support(self, *args, **kwargs):
+        return False
 
     class Meta:
         model = models.CommunicationsTopic
@@ -92,6 +124,10 @@ class MessageView(
     @action(methods=["patch"], detail=True)
     def read(self, request, pk):
         my_entity = models.entity_for_current_user(self.request)
+
+        if not my_entity:  # pragma: no cover
+            raise ValidationError("Invalid entity to mark message as read")
+
         obj = self.get_object()
 
         if not obj.sent_at:  # pragma: no cover
@@ -103,13 +139,22 @@ class MessageView(
     @action(methods=["patch"], detail=True)
     def unread(self, request, pk):
         my_entity = models.entity_for_current_user(self.request)
+
+        if not my_entity:  # pragma: no cover
+            raise ValidationError("Invalid entity to mark message as unread")
+
         obj = self.get_object()
         obj.read_by.filter(entity=my_entity).delete()
         return self.retrieve(request, pk)
 
-    def _annotate_read_flag(self, qs):
-        my_entity = models.entity_for_current_user(self.request)
-        my_read = models.CommunicationsReadMarker.objects.all().filter(entity=my_entity)
+    def _annotate_read_flag(self, qs, entity):
+        """Annotate the given queryset to have a `read_at` attribute.
+
+        This gives information as to when the given entity (usually current user)
+        has read a given message in the queryset.
+        """
+
+        my_read = models.CommunicationsReadMarker.objects.all().filter(entity=entity)
 
         qs = qs.annotate(
             read_at=my_read.filter(message=OuterRef("pk"))
@@ -122,8 +167,34 @@ class MessageView(
         qs = super().get_queryset(*args, **kwargs)
         my_entity = models.entity_for_current_user(self.request)
 
-        qs = qs.filter(topic__involved_entities__contains=[my_entity])
-        return self._annotate_read_flag(qs)
+        if not my_entity:  # pragma: no cover
+            # Just a safeguard. Will never happen because of permissions
+            return qs.none()
+
+        qs = self._annotate_read_flag(qs, my_entity)
+
+        return qs.filter(topic__involved_entities__contains=[my_entity])
+
+    @permission_aware
+    def has_object_read_permission(self, *args, **kwargs):
+        return True
+
+    def has_object_read_permission_for_support(self, *args, **kwargs):
+        return False
+
+    @permission_aware
+    def has_object_unread_permission(self, *args, **kwargs):
+        return True
+
+    def has_object_unread_permission_for_support(self, *args, **kwargs):
+        return False
+
+    @permission_aware
+    def has_create_permission(self, *args, **kwargs):
+        return True
+
+    def has_create_permission_for_support(self, *args, **kwargs):
+        return False
 
     class Meta:
         model = models.CommunicationsMessage
@@ -140,6 +211,7 @@ class AttachmentView(
     # DRF
     RetrieveModelMixin,
     ListModelMixin,
+    DestroyModelMixin,
     GenericViewSet,
 ):
     serializer_class = serializers.CommunicationsAttachmentSerializer
@@ -148,8 +220,16 @@ class AttachmentView(
     queryset = models.CommunicationsAttachment.objects
 
     def get_queryset(self, *args, **kwargs):
-        my_entity = models.entity_for_current_user(self.request)
         qs = super().get_queryset(*args, **kwargs)
+
+        if is_support(self.request):
+            return qs
+
+        my_entity = models.entity_for_current_user(self.request)
+
+        if not my_entity:  # pragma: no cover
+            # Just a safeguard. will never happen in production
+            return qs.none()
 
         qs = qs.filter(message__topic__involved_entities__contains=[my_entity])
         return qs
@@ -168,15 +248,22 @@ class AttachmentView(
         serializer.save()
         return response.Response(serializer.data)
 
+    @permission_aware
     def has_object_convert_to_document_permission(self, attachment):
         """Check if user has permission to convert a given attachment to a document."""
-
         my_entity = models.entity_for_current_user(self.request)
-        if my_entity == "APPLICANT":
+
+        if not my_entity:  # pragma: no cover
             return False
 
         involved_entities = attachment.message.topic.involved_entities
         return my_entity in involved_entities
+
+    def has_object_convert_to_document_permission_for_support(self, attachment):
+        return False
+
+    def has_object_convert_to_document_permission_for_applicant(self, attachment):
+        return False
 
     @action(methods=["get"], detail=True, permission_classes=[])
     def download(self, request, pk=None):
@@ -208,6 +295,13 @@ class AttachmentView(
             as_attachment=as_attachment,
             content_type=obj.content_type,
         )
+
+    @permission_aware
+    def has_object_destroy_permission(self, attachment):
+        return False
+
+    def has_object_destroy_permission_for_support(self, attachment):
+        return True
 
     class Meta:
         model = models.CommunicationsAttachment
